+++ android/src/test/kotlin/com/adyen/checkout/flutter/ConfigurationMapperTest.kt
@@ -46,6 +46,7 @@ import com.adyen.checkout.googlepay.GooglePayConfiguration
 import com.adyen.checkout.twint.TwintConfiguration
 import com.google.android.gms.wallet.WalletConstants
 import org.junit.jupiter.api.Assertions.assertEquals
+import org.junit.jupiter.api.Assertions.assertNotNull
 import org.junit.jupiter.api.Assertions.assertNull
 import org.junit.jupiter.api.Nested
 import org.junit.jupiter.api.Test
@@ -582,6 +583,7 @@ class ConfigurationMapperTest {
                 shopperLocale = "nl-NL",
                 amount = AmountDTO("EUR", 5000),
                 analyticsOptionsDTO = AnalyticsOptionsDTO(true, "1.0.0"),
+                threeDS2ConfigurationDTO = null,
             )
 
             val checkoutConfiguration = actionConfigurationDTO.toCheckoutConfiguration()
@@ -591,6 +593,7 @@ class ConfigurationMapperTest {
             assertEquals("nl-NL", checkoutConfiguration.shopperLocale?.toLanguageTag())
             assertEquals("EUR", checkoutConfiguration.amount?.currency)
             assertEquals(5000, checkoutConfiguration.amount?.value)
+            assertNull(checkoutConfiguration.getActionConfiguration(Adyen3DS2Configuration::class.java))
         }
 
         @Test
@@ -601,6 +604,7 @@ class ConfigurationMapperTest {
                 shopperLocale = null,
                 amount = null,
                 analyticsOptionsDTO = AnalyticsOptionsDTO(false, "1.0.0"),
+                threeDS2ConfigurationDTO = null,
             )
 
             val checkoutConfiguration = actionConfigurationDTO.toCheckoutConfiguration()
@@ -608,6 +612,27 @@ class ConfigurationMapperTest {
             assertEquals(SDKEnvironment.UNITED_STATES, checkoutConfiguration.environment)
             assertNull(checkoutConfiguration.shopperLocale)
             assertNull(checkoutConfiguration.amount)
+            assertNull(checkoutConfiguration.getActionConfiguration(Adyen3DS2Configuration::class.java))
+        }
+
+        @Test
+        fun `when action component configuration has 3DS2 configuration, then map correctly`() {
+            val actionConfigurationDTO = ActionComponentConfigurationDTO(
+                environment = Environment.TEST,
+                clientKey = TEST_CLIENT_KEY,
+                shopperLocale = null,
+                amount = null,
+                analyticsOptionsDTO = AnalyticsOptionsDTO(true, "1.0.0"),
+                threeDS2ConfigurationDTO = ThreeDS2ConfigurationDTO(
+                    requestorAppURL = "https://example.com/action-3ds2"
+                ),
+            )
+
+            val checkoutConfiguration = actionConfigurationDTO.toCheckoutConfiguration()
+            val threeDS2Configuration = checkoutConfiguration.getActionConfiguration(Adyen3DS2Configuration::class.java)
+
+            assertNotNull(threeDS2Configuration)
+            assertEquals("https://example.com/action-3ds2", threeDS2Configuration?.threeDSRequestorAppURL)
         }
     }