+++ pigeons/platform_api.dart
@@ -777,6 +777,7 @@ class ActionComponentConfigurationDTO {
   final String? shopperLocale;
   final AmountDTO? amount;
   final AnalyticsOptionsDTO analyticsOptionsDTO;
+  final ThreeDS2ConfigurationDTO? threeDS2ConfigurationDTO;
 
   ActionComponentConfigurationDTO(
     this.environment,
@@ -784,6 +785,7 @@ class ActionComponentConfigurationDTO {
     this.shopperLocale,
     this.amount,
     this.analyticsOptionsDTO,
+    this.threeDS2ConfigurationDTO,
   );
 }